Move EventController and UserController to web package for layer separation
