+++ src/main/java/edu/uc/campusevent/domain/event/EventCategory.java
